# public/openapi.yml
openapi: 3.1.0
servers:
  - url: https://inspector.duckarchive.com
    description: Production server
info:
  title: inspector.duckarchive.com API
  version: 2.0.0
  description: >
    AI agents and other automated clients: see /llms.txt for an up-to-date,
    agent-oriented guide to the public read-only endpoints (/api/search,
    /api/catalog, /api/catalog/**, /api/authors), including a rate-limit
    note (5 requests per 10 seconds per client).


    Editor/moderation endpoints under /api/editor/** exist but require an
    authenticated NextAuth session (cookie-based, not the bearer token below)
    and are not documented here — they aren't part of the public API surface.
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: >
        A Duck account token (Authorization: Bearer <token>), unrelated to the
        NextAuth session cookies used by /api/auth/** and the editor UI.
  schemas:
    ErrorMessage:
      type: object
      properties:
        message:
          type: string
    YearRange:
      type: object
      properties:
        start_year:
          type: integer
        end_year:
          type: integer
    Archive:
      type: object
      properties:
        id:
          type: string
          format: uuid
        code:
          type: string
        updated_at:
          type: string
          format: date-time
          nullable: true
        title:
          type: string
          nullable: true
        info:
          type: string
          nullable: true
        url:
          type: string
          nullable: true
        address:
          type: string
          nullable: true
        phone_number:
          type: string
          nullable: true
        email:
          type: string
          nullable: true
        logo_url:
          type: string
          nullable: true
    FondSummary:
      type: object
      properties:
        id:
          type: string
          format: uuid
        code:
          type: string
        title:
          type: string
          nullable: true
        years:
          type: array
          items:
            $ref: "#/components/schemas/YearRange"
    ArchiveWithFonds:
      allOf:
        - $ref: "#/components/schemas/Archive"
        - type: object
          properties:
            fonds:
              type: array
              items:
                $ref: "#/components/schemas/FondSummary"
...
    OnlineCopy:
      type: object
      properties:
        id:
          type: string
          format: uuid
        updated_at:
          type: string
          format: date-time
        checked_availability_at:
          type: string
          format: date-time
          nullable: true
        resource_id:
          type: string
          format: uuid
        inventory_id:
          type: string
          format: uuid
          nullable: true
        file_id:
          type: string
          format: uuid
          nullable: true
        url:
          type: string
        availability:
          type: string
          enum: [PUBLIC, RESTRICTED]
          nullable: true
        parsed:
          type: string
...
    Author:
      type: object
      properties:
        id:
          type: string
          format: uuid
        title:
          type: string
        info:
          type: string
          nullable: true
        lat:
          type: number
          nullable: true
        lng:
          type: number
          nullable: true
        tags:
          type: array
          items:
            type: string
    FileLocation:
      type: object
      properties:
        id:
          type: string
          format: uuid
        lat:
          type: number
        lng:
          type: number
        radius_m:
          type: integer
    File:
      type: object
      properties:
        id:
          type: string
          format: uuid
        code:
          type: string
        full_code:
          type: string
        updated_at:
          type: string
          format: date-time
          nullable: true
        title:
          type: string
          nullable: true
        info:
          type: string
          nullable: true
        tags:
          type: array
          items:
            type: string
        inventory_id:
          type: string
          format: uuid
    FileWithDetails:
      allOf:
        - $ref: "#/components/schemas/File"
        - type: object
          properties:
            years:
              type: array
              items:
                $ref: "#/components/schemas/YearRange"
            online_copies:
              type: array
              items:
                $ref: "#/components/schemas/OnlineCopy"
            authors:
              type: array
              items:
                type: object
                properties:
                  file_id:
                    type: string
                    format: uuid
                  author_id:
                    type: string
                    format: uuid
                  author:
                    $ref: "#/components/schemas/Author"
            locations:
              type: array
              items:
                $ref: "#/components/schemas/FileLocation"
            inventory:
              type: object
              properties:
                fond_id:
                  type: string
                  format: uuid
    SearchResultFile:
      type: object
      properties:
        id:
          type: string
          format: uuid
        code:
          type: string
        updated_at:
          type: string
          format: date-time
        title:
          type: string
        info:
          type: string
          nullable: true
        inventory_id:
          type: string
          format: uuid
        description_id:
          type: string
          format: uuid
          description: Legacy alias of inventory_id, kept for existing API consumers.
        full_code:
          type: string
        tags:
          type: array
          items:
            type: string
        is_online:
          type: boolean
        years:
          type: array
          items:
            type: object
            properties:
              file_id:
                type: string
                format: uuid
              case_id:
                type: string
                format: uuid
                description: Legacy alias of file_id.
              start_year:
                type: integer
              end_year:
                type: integer
paths:
  /api/search:
    post:
      summary: Search files
      description: >
        Search across all files by title, author, place, geo-radius, year
        range, tags, or exact catalog codes. All body fields are optional and
        combined with AND. Results are capped at 50 rows with no further
        pagination — narrow the query instead. Rate limit: 5 requests per 10
        seconds per client (see /llms.txt).
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                title:
                  type: string
                  description: >
                    Free-text match on the file's title and info and on the
                    name/info of its linked authors (church, parish).
                fuzziness:
                  type: number
                  minimum: 0
                  maximum: 1
                  description: >
                    Optional trigram word-similarity threshold applied to title,
                    place and author (1 = exact words, 0.3 = very loose; lower values are raised to 0.3). Absent
                    or 0 = plain substring match.
                author:
                  type: string
                  description: Substring match on a linked author's name.
                place:
                  type: string
                  description: >
                    Substring match on file/author location text. Mutually
                    exclusive with lat/lng/radius_m.
                lat:
                  type: string
                  description: Latitude for geo-radius search (mutually exclusive with place).
                lng:
                  type: string
                  description: Longitude for geo-radius search (mutually exclusive with place).
                radius_m:
                  type: integer
                  description: Search radius in meters, used with lat/lng.
                year_from:
                  type: string
                  description: Start of the year range (inclusive, overlap match).
                year_to:
                  type: string
                  description: End of the year range (inclusive, overlap match).
                tags:
                  type: array
                  items:
                    type: string
                  description: Matched against file or author tags.
                archive:
                  type: string
                  description: Exact archive code (case-sensitive).
                fond:
                  type: string
                  description: Exact fond code (case-sensitive).
                inventory:
                  type: string
                  description: Exact inventory code (case-sensitive).
                file:
                  type: string
                  description: Exact file code (case-sensitive).
                is_online:
                  type: boolean
                  description: Only files with a publicly available online copy.
      responses:
        "400":
          description: >
            Invalid input — any Latin letter in a text value, an unknown field,
            a non-numeric year/coordinate. The public API only accepts Cyrillic
            catalog data.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorMessage"
        "200":
          description: Up to 50 matching files.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/SearchResultFile"
        "500":
          description: Search failed.
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
  /api/authors:
    get:
      summary: Search authors
      description: Returns up to 200 authors, optionally filtered by name.
      parameters:
        - name: q
          in: query
          required: false
          schema:
            type: string
          description: Substring match on author title/name.
      responses:
        "200":
          description: List of authors.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Author"
...
  /api/catalog/{archive-code}:
    get:
      summary: Get archive and its fonds
      parameters:
        - name: archive-code
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Archive with its fonds.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ArchiveWithFonds"
        "400":
          description: Missing archive-code.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorMessage"
        "404":
          description: Archive not found.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorMessage"
...
  /api/catalog/{archive-code}/{fond-code}/{inventory-code}/{file-code}:
    get:
      summary: Get a single file
      description: Returns the file with its years, authors, geo locations, and online copies.
      parameters:
        - name: archive-code
          in: path
          required: true
          schema:
            type: string
        - name: fond-code
          in: path
          required: true
          schema:
            type: string
        - name: inventory-code
          in: path
          required: true
          schema:
            type: string
        - name: file-code
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: File with full details.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FileWithDetails"
        "400":
          description: Missing path parameter.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorMessage"
        "404":
          description: File not found.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorMessage"
...